set a loading state for loading kitchens. also dropped/raised the kithcnesImInContext to app.tsx too so on logout I can clear all contexts
